+++ app/basic.rb
@@ -1,3 +1,8 @@
+# frozen_string_literal: true
+
 require_relative '../lib/basic'
 require_relative '../lib/flatplan'
+require_relative 'tool'
 include Flatplan
+
+APP_NAME = 'FlatPlan'