+++ lib/flatplan/core.rb
@@ -5,9 +5,9 @@ require_relative 'core/media_assets'
 require_relative 'core/text_and_media'
 require_relative 'core/story'
 require_relative 'core/factory'
-require_relative 'core/story_initializer'
-require_relative 'core/story_serializer'
-require_relative 'core/story_builder'
+require_relative 'core/initializer'
+require_relative 'core/serializer'
+require_relative 'core/builder'
 
 module Flatplan
   # Content model namespace